refactor: consolidate build configuration into app.config.ts
- Remove build.config.js and integrate constants directly into app.config.ts
- Add BUILD_CONFIG constant with all build-time values
- Implement conditional Universal Links configuration (empty by default)
- Move runtime constants (store URLs, social links) to Config.ts with proper defaults
- Simplify ExtraConfig interface to only pass build-time computed values
- Add BUILD_TIME field for tracking build timestamp
- Update documentation to reflect new configuration structure
