+++ docs/configuration-guide.md
@@ -3,7 +3,8 @@
 ## Configuration System
 
 ```
-config.ts          →  App constants (Bundle ID, Scheme, Store URLs)
+app.config.ts      →  Build-time constants and Expo configuration
+Config.ts          →  Runtime constants with defaults (Store URLs, Social Links, etc.)
 .env files         →  Local development only (NOT used in EAS builds)
 eas.json           →  Build profiles and non-sensitive build variables
 EAS Cloud          →  Sensitive variables (API keys, secrets)
@@ -11,12 +12,12 @@ EAS Cloud          →  Sensitive variables (API keys, secrets)
 
 ## Key Files
 
-### 1. `build.config.js`
-Centralized build configuration - edit this file to change:
-- App identity (Bundle ID, Scheme, Slug, Owner)
-- Store URLs and Social Media Links  
-- Legal information (Privacy, Terms, Copyright)
-- Website and Support Email
+### 1. `app.config.ts`
+Build-time configuration - contains:
+- BUILD_CONFIG constant with app identity (Bundle ID, Scheme, Slug, Owner)
+- Universal Links configuration (empty by default)
+- Environment-specific logic for dev/preview/production
+- Passes configuration to runtime via extra field
 
 ### 2. `src/constants/Config.ts`
 Runtime configuration access - use this in your app:
@@ -54,6 +55,34 @@ Priority: `.env.local` > `.env` > hardcoded defaults
 2. `eas.json` env field - for non-sensitive config
 3. ❌ Local .env files are NOT used
 
+## Adding New Configuration
+
+### 1. Runtime constant (e.g., Facebook URL):
+```typescript
+// In src/constants/Config.ts only
+FACEBOOK_URL: 'https://facebook.com/buildeasy',
+```
+
+### 2. Environment variable (including sensitive):
+```typescript
+// Step 1: For sensitive values, create EAS secret
+eas secret:create EXPO_PUBLIC_STRIPE_KEY pk_test_xxx
+
+// Step 2: Add to app.config.ts extra
+extra: {
+  stripeKey: process.env.EXPO_PUBLIC_STRIPE_KEY || '',
+}
+
+// Step 3: Add to ExtraConfig interface and Config.ts
+STRIPE_KEY: extra?.stripeKey || '',
+```
+
+### 3. Build-time constant:
+```typescript
+// In app.config.ts BUILD_CONFIG
+NEW_BUILD_CONSTANT: 'value',
+```
+
 ## Common Tasks
 
 **Local setup:**
@@ -80,19 +109,17 @@ eas build --profile production
 import { Config } from '@/constants/Config';
 const bundleId = Config.BUNDLE_ID;
 const supportEmail = Config.SUPPORT_EMAIL;
-
-// Environment variables (direct access)
-const apiUrl = process.env.EXPO_PUBLIC_API_URL;
+const apiUrl = Config.API_URL; // Environment variables accessed via Config
 ```
 
 ## Configuration Strategy
 
 ### Where to put configurations:
 
-**`build.config.js`** - Build constants
+**`app.config.ts` BUILD_CONFIG** - Build constants
 - Bundle ID, URL schemes, Slug
-- Store URLs, Social Media, Legal info
-- Things that rarely change across environments
+- Universal Links domains
+- Build numbers
 
 **`.env` files** - Local development only
 - API URLs for local testing
